XEOK-272: Remove skybox from the viewer heirarchy
