+++ examples/viewer/index.html
@@ -272,6 +272,10 @@
             ["logDepthBuf_ortho", "Enabling the logarithmic depth buffer, with orthographic camera"]
         ],
 
+        "Skybox": [
+            ["skybox.html", "Displaying a skybox in the background"],
+        ],
+
         "Spinner": [
             "#CSS animation while models are loading",
             ["spinner_custom", "Customizing the appearance of the loading spinner"]
