+++ examples/viewer/skybox.html
@@ -39,16 +39,12 @@
     <canvas id="myCanvas"></canvas>
     
     <div class="slideout-sidebar">
-        <h1>SphereMaps</h1>
-        <h2>Click on the images on the left side to switch between SphereMaps</h2>
-        <p>In this example, we are rendering a 360 degree equirectangular image.</p>
+        <h1>Skybox</h1>
+        <p>In this example, we are rendering a skybox.</p>
         <h3>Components Used</h3>
         <ul>
             <li>
                 <a href="../../docs/class/src/viewer/Viewer.js~Viewer.html" target="_other">Viewer</a>
-            </li>
-            <li>
-                <a href="../../docs/class/src/viewer/scene/skybox/Skybox.js~Skybox.html" target="_other">Skybox</a>
             </li>
             <li>
                 <a href="../../docs/class/src/plugins/XKTLoaderPlugin/XKTLoaderPlugin.js~XKTLoaderPlugin.html"
@@ -70,7 +66,7 @@
     // Import the modules we need for this example
     //------------------------------------------------------------------------------------------------------------------
 
-    import { Viewer, Skybox, XKTLoaderPlugin } from "../../dist/xeokit-sdk.min.es.js";
+    import { Viewer, createSkyboxMesh, createCombinedTexture, XKTLoaderPlugin } from "../../dist/xeokit-sdk.min.es.js";
 
     //------------------------------------------------------------------------------------------------------------------
     // Create a Viewer and arrange the camera
@@ -89,16 +85,25 @@
     // Add a skybox image
     //------------------------------------------------------------------------------------------------------------------
 
-    new Skybox(viewer.scene, {
+    try {
+        
+        const texture = await createCombinedTexture(viewer.scene, {
         src: [
-            "../../assets/images/san-francisco-posx.jpg",
-            "../../assets/images/san-francisco-negx.jpg",
-            "../../assets/images/san-francisco-posy.jpg",
-            "../../assets/images/san-francisco-negy.jpg",
-            "../../assets/images/san-francisco-posz.jpg",
-            "../../assets/images/san-francisco-negz.jpg"
-        ],
-    })
+                "../../assets/images/san-francisco-posx.jpg",
+                "../../assets/images/san-francisco-negx.jpg",
+                "../../assets/images/san-francisco-posy.jpg",
+                "../../assets/images/san-francisco-negy.jpg",
+                "../../assets/images/san-francisco-posz.jpg",
+                "../../assets/images/san-francisco-negz.jpg"
+            ]
+        })
+
+        const skyboxMesh = createSkyboxMesh(viewer.scene, texture);
+        skyboxMesh.scale = [2000, 2000, 2000];
+
+    } catch(e) {
+        console.log('error at creating skybox: ', e);
+    }
 
     //------------------------------------------------------------------------------------------------------------------
     // Load an xkt model
