+++ src/extras/index.js
@@ -3,3 +3,4 @@ export * from "./PointerLens/index.js";
 export * from "./collision/index.js";
 export * from "./MarqueePicker/index.js";
 export * from "./PointerCircle/index.js";
+export * from "./Skybox/index.js";
