--- src/plugins/SkyboxesPlugin/index.js
@@ -1,1 +1,0 @@
-export * from "./SkyboxesPlugin.js";
