+++ src/plugins/index.js
@@ -10,7 +10,6 @@ export * from "./OBJLoaderPlugin/index.js";
 export * from "./SectionPlanesPlugin/index.js";
 export * from "./StoreyViewsPlugin/index.js";
 export * from "./FaceAlignedSectionPlanesPlugin/index.js";
-export * from "./SkyboxesPlugin/index.js";
 export * from "./STLLoaderPlugin/index.js";
 export * from "./TreeViewPlugin/index.js";
 export * from "./ViewCullPlugin/index.js";
