+++ src/viewer/scene/index.js
@@ -13,7 +13,6 @@ export * from "./nodes/index.js";
 export * from "./paths/index.js";
 export * from "./model/index.js";
 export * from "./sectionPlane/index.js";
-export * from "./skybox/index.js";
 export * from "./utils/index.js";
 export * from "./Component.js";
 export * from "./utils.js";
