+++ types/viewer/scene/index.d.ts
@@ -16,7 +16,6 @@ export * from './paths';
 export * from "./models";
 export * from "./scene/Scene";
 export * from "./sectionPlane";
-export * from './skybox';
 export * from "./stats";
 export * from "./viewport";
 export * from "./webgl";
