--- types/viewer/scene/skybox/index.d.ts
@@ -1,1 +1,0 @@
-export * from './Skybox';
